+++ docs/changelog.rst
@@ -17,6 +17,7 @@ Fixed
 
 - Pytest fixture parametrization via ``pytest_generate_tests``. `#280`_
 - Support for tests defined as methods. `#282`_
+- Unclosed ``requests.Session`` on calling ``Case.call`` without passing a session explicitly. `#286`_
 
 `0.15.0`_ - 2019-11-15
 ----------------------
@@ -414,6 +415,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#286: https://github.com/kiwicom/schemathesis/issues/286
 .. _#282: https://github.com/kiwicom/schemathesis/issues/282
 .. _#280: https://github.com/kiwicom/schemathesis/issues/280
 .. _#272: https://github.com/kiwicom/schemathesis/issues/272